* Working on Networking, specifically connections, connection authorisation and   data initialisation. Have tested each in a test application (using   NetworkController to make a connection), and everything up to and including   initialisation seems to work (data sent correctly, and data integrity remains)

+++ 2DPhysicsDemo/ThreadPool.cpp
@@ -40,7 +40,7 @@ bool ThreadPool::Alive()
 
 int ThreadPool::InitPool(int threadCount)
 {
-	if(threads.size() == threadCount) return -1; // already initialised, dont reinit
+	if(threads.size() == (unsigned int)threadCount) return -1; // already initialised, dont reinit
 
 	// otherwise there is < or > the thread count we want, so kill all the current threads (if any) and reset
 	SigKill();
